CLN: clean-up internal sparse imports + restructure sparse submodule (#28516)
