+++ pandas/tests/arrays/sparse/test_arithmetics.py
@@ -5,7 +5,7 @@ import pytest
 
 import pandas as pd
 from pandas.core import ops
-from pandas.core.sparse.api import SparseDtype
+from pandas.core.arrays.sparse import SparseDtype
 import pandas.util.testing as tm
