+++ pandas/tests/reshape/test_reshape.py
@@ -8,7 +8,7 @@ from pandas.core.dtypes.common import is_integer_dtype
 
 import pandas as pd
 from pandas import Categorical, DataFrame, Index, Series, get_dummies
-from pandas.core.sparse.api import SparseArray, SparseDtype
+from pandas.core.arrays.sparse import SparseArray, SparseDtype
 import pandas.util.testing as tm
 from pandas.util.testing import assert_frame_equal
